+++ .env
@@ -3,6 +3,8 @@ DB_USER=root
 DB_PASSWORD=
 DB_PORT=3306
 
+SECRET="Alkemy"
+
 AWS_ID="AKIA56D2VOAWK2T27SY6"
 AWS_SECRET="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"
 AWS_BUCKET_NAME="ong-bucket-alkemy"
