--- .env.example
@@ -1,4 +1,0 @@
-DB_NAME=ong
-DB_USER=root
-DB_PASSWORD=
-DB_PORT=3306
