+++ package.json
@@ -16,6 +16,7 @@
     "express": "~4.16.1",
     "express-validator": "^6.14.0",
     "http-errors": "~1.6.3",
+    "jsonwebtoken": "^8.5.1",
     "morgan": "~1.9.1",
     "multer": "^1.4.4",
     "mysql2": "^2.2.3",
