+++ utils/s3.js
@@ -1,0 +1,10 @@
+require('dotenv').config()
+const AWS = require('aws-sdk')
+
+
+const s3 = new AWS.S3({
+    accessKeyId: process.env.AWS_ID,
+    secretAccessKey: process.env.AWS_SECRET
+})
+
+module.exports = s3
